Decouple Storm & Cache logic. Create sender interface. Move topology related classes to transport package. Add unit testing for Cache bolt.
