+++ services/wfm/app/Makefile
@@ -5,7 +5,7 @@ TOPOLOGY := wfm flow stats cache islstats opentsdb portstate nbworker ping
 ENTRY_wfm := org.openkilda.wfm.topology.event.OfEventWfmTopology
 ENTRY_flow := org.openkilda.wfm.topology.flow.FlowTopology
 ENTRY_stats := org.openkilda.wfm.topology.stats.StatsTopology
-ENTRY_cache := org.openkilda.wfm.topology.cache.CacheTopology
+ENTRY_cache := org.openkilda.wfm.topology.cache.transport.CacheTopology
 ENTRY_islstats := org.openkilda.wfm.topology.islstats.IslStatsTopology
 ENTRY_opentsdb := org.openkilda.wfm.topology.opentsdb.OpenTSDBTopology
 ENTRY_portstate := org.openkilda.wfm.topology.portstate.PortStateTopology
