+++ services/wfm/src/main/java/org/openkilda/wfm/topology/cache/StreamType.java
@@ -15,6 +15,8 @@
 
 package org.openkilda.wfm.topology.cache;
 
+import org.openkilda.wfm.topology.cache.transport.CacheTopology;
+
 /**
  * Represents stream used in {@link CacheTopology}.
  */
