+++ services/wfm/src/main/java/org/openkilda/wfm/topology/cache/transport/CacheTopology.java
@@ -13,7 +13,7 @@
  *   limitations under the License.
  */
 
-package org.openkilda.wfm.topology.cache;
+package org.openkilda.wfm.topology.cache.transport;
 
 import org.openkilda.messaging.ServiceType;
 import org.openkilda.pce.provider.Auth;
@@ -23,6 +23,7 @@ import org.openkilda.wfm.LaunchEnvironment;
 import org.openkilda.wfm.config.Neo4jConfig;
 import org.openkilda.wfm.error.NameCollisionException;
 import org.openkilda.wfm.topology.AbstractTopology;
+import org.openkilda.wfm.topology.cache.StreamType;
 
 import org.apache.storm.generated.ComponentObject;
 import org.apache.storm.generated.StormTopology;
@@ -88,10 +89,9 @@ public class CacheTopology extends AbstractTopology<CacheTopologyConfig> {
         CacheBolt cacheBolt = new CacheBolt(pathComputerAuth);
         ComponentObject.serialized_java(org.apache.storm.utils.Utils.javaSerialize(pathComputerAuth));
         BoltDeclarer boltSetup = builder.setBolt(BOLT_ID_CACHE, cacheBolt, parallelism)
-                .shuffleGrouping(SPOUT_ID_COMMON)
+                .shuffleGrouping(SPOUT_ID_COMMON);
         // (carmine) as per above comment, only a single input streamt
         //                .shuffleGrouping(SPOUT_ID_TOPOLOGY)
-        ;
         ctrlTargets.add(new CtrlBoltRef(BOLT_ID_CACHE, cacheBolt, boltSetup));
 
         /*
